Enable bloom filter on tables
This currently reaches under the hood of the connection to the client
itself to call the setTableProperty.

self.conn.client.setTableProperty(self.conn.login, self._ns(table),
                                  'table.bloom.enabled', 'true')

We should contribute back to the pyaccumulo project a set of functions
for managing the table properties.  This would be quite easy to do.
Add them to Accumulo class in pyaccumulo/__init__.py