Add retry decorator
A decorator that will retry a function a limited number of
times upon a configurable class of failures.

Initially this decorator is only being applied to the put method of
accumulo.  We are applying it here because we observed failures at
points when accumulo itself hadn't actually failed.  Accumulo
occassionally takes a long time to respond to certain requests.
(Probably during periods where compactions are taking place.) The
downside to this retry is that sometimes accumulo does experience an
error where a retry will not be helpful.  Unfortunately the same error
appears in both cases.

More generally, I am hesitant to apply the retry decorator elsewhere
unless we have a good understanding of the causes of failures we are
retrying on.  It would be best to tune backends and clients to avoid
most failures.