pycassa >= 1.10
"Batches are atomic by default starting in 1.2. Unfortunately, the
price for atomicity is about a 30% hit to performance compared to the
old non-atomic batches. So we also provide BEGIN UNLOGGED BATCH for
when performance is more important than atomicity guarantees."

The latest version of pycassa (released Oct 3) has support for atomic
update and has it default to false.