If accumulo has multiple storage_addresses, pick one randomly.
This is a change from the previous "pick the first one" policy, and it
means that if there are multiple Accumulo proxies in the cluster, traffic
should hit all of them instead of concentrating on just one.  So a single
client object will still only talk to a single proxy, but multiple clients
(maybe in multiple processes) will spread work around.