kvlayer can discover (postgrest) storage from environment variables.
There is a general hook for other backends to provide discovery too,
but right now only postgrest actually does it.  If there are
POSTGRES_PORT_5432_TCP_ADDR and POSTGRES_PORT_5432_TCP_PORT environment
variables, and there is not storage_type and/or storage_addresses, then
postgrest is used as the storage backend with a generated
storage_addresses pointing at that.  This is intended to be used with
the Docker postgres:9.3 image --link postgres:postgres into the running
container.