Accumulo: use SummingCombiner for COUNTER; delete properly.
If a table has COUNTER type, attach a SummingCombiner to it.  Then
values are 8-byte big-endian longs, and basic cell writes in effect
add to the cell value.  Special-case code for encoding, increment, and
put (including bringing back **kwargs on the StringKeyedStorage
interface functions, since we need a backend-specific kwarg to do this).

Don't use RowDeletingIterator, just delete cells we're going to delete
and be done with it.  (RowDeletingIterator is mildly more interesting
if you have more than one cell per row.)