Trim the sign bit of the snowflake key.
The "packed" encoder assumes a 64-bit *signed* long and asserts that
its value is in range for this, but bit packing is much easier to do on
*unsigned* longs.  Give in and just trim the time range to 31 unsigned
bits.

(To consider: use 10 bits for a fractional second.  Then the useful range
of a key would be 21 bits...which is only 24 days, sort of impractically
short.)