Add connection pooling + recovery to Hbase backend.
This takes the same approach as pooling+recovery in the `postgrest`
backend. In particular, it uses a context manager to create new
connections. Before handing out the connection, it does a trivial
check to make sure the connection is healthy. If it isn't, it tries
to refresh the underlying socket.

We cap retries to 5 at a 0.5 second interval. This could be more clever
(perhaps exponential backoff).

Finally, this is completely untested.